requirements.txt
PyQt6
pyperclip
pyttsx3
keyboard
Pillow
opencv-python
numpy
opencc-python-reimplemented
paddleocr>=3.0.0
